services/polymarket-agent/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
google-cloud-firestore==2.19.0
httpx==0.27.0
openai==1.51.0
tavily-python==0.5.0
numpy==1.26.0
python-dotenv==1.0.0
websockets==12.0
web3==6.20.0
